Release 1, fixing bugs version 7

+++ app/controllers/StationCtrl.java
@@ -18,10 +18,12 @@ public class StationCtrl extends Controller
         Station station = Station.findById(id);
         Reading reading = Reading.findById(readingid);
         Logger.info ("Removing" + reading.id);
+        if (reading != null) {
         station.readings.remove(reading);
         station.save();
         reading.delete();
         render("station.html", station);
+        }
     }
     public static void addReading(Long id, int code, float temperature, float windSpeed, int pressure, float windDirection)
     {
